Todas Manutenções estruturadas, falta busca e alguns delete

+++ Cinema/src/java/com/br/lp2/cinema/commands/DeletaSala.java
@@ -2,6 +2,9 @@ package com.br.lp2.cinema.commands;
 
 import com.br.lp2.cinema.model.DAO.SalaDAO;
 import com.br.lp2.cinema.model.DAO.SalaDAOconcreto;
+import com.br.lp2.cinema.model.DAO.SessaoDAO;
+import com.br.lp2.cinema.model.DAO.SessaoDAOconcreto;
+import com.br.lp2.cinema.model.javabeans.Sessao;
 import java.io.IOException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -17,16 +20,28 @@ public class DeletaSala implements Command{
     public void execute(HttpServletRequest request, HttpServletResponse response) {
         codigo = Integer.parseInt(request.getParameter("codigo"));
         
-        SalaDAO salaDAO = new SalaDAOconcreto();
-        boolean delete = salaDAO.deleteSala(codigo);
+        SessaoDAO sessaoDAO = new SessaoDAOconcreto();
+        Sessao sessao = sessaoDAO.readSessaoBySala(codigo);
         
-        try {
-            if(delete)
-                response.sendRedirect("sucesso.html");
-            else
+        if(sessao == null){
+            SalaDAO salaDAO = new SalaDAOconcreto();
+            boolean delete = salaDAO.deleteSala(codigo);
+        
+            try {
+                if(delete)
+                    response.sendRedirect("sucesso.html");
+                else
+                    response.sendRedirect("principal_gerente.jsp");
+            } catch (IOException ex) {
+                ex.getMessage();
+            }
+        }
+        else{
+            try{
                 response.sendRedirect("manter_sala.jsp");
-        } catch (IOException ex) {
-            ex.getMessage();
+            } catch (IOException ex){
+                ex.getMessage();
+            }
         }
     }
 }
